refactor(124122): :hammer: refatora socket
